requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
python-dotenv==1.0.0
langchain==0.1.0
openai==1.8.0
anthropic==0.8.1
groq==0.4.1
google-cloud-aiplatform==1.38.1
cohere==4.39
pydantic==2.5.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2
loguru==0.7.2
pandas==2.1.4
numpy==1.26.3
pytest==7.4.4
pyjwt==2.3.0
